Moved changes out of plan_solver.py
